design fixes
### <BaseFilterNav>
- added an opotional title prop to to help seperate on home what section a user is watching right now.
( this was planed in layout by say "ACTIVE ROUNDS" not just "ACTIVE",
but i think as its now works fine as well.

- fixed active vs not active class styles for a tab to match the layout
( inactives are grey ,  actives are not underline but border-b )

think i need to rework this later on for better mobile experience but its good for now.

### Home.vue
- fixed the empty state
( had no bottom-line and needed some vertical padding )

### LoadingSpinner.vue
- rewrote a bit the css for responsive spinner image